+++ ubo_app/colors.py
@@ -1,0 +1,26 @@
+"""Defines the color constants used in the application."""
+
+from ubo_gui.constants import (
+    DANGER_COLOR,
+    INFO_COLOR,
+    PRIMARY_COLOR,
+    SECONDARY_COLOR,
+    SECONDARY_COLOR_LIGHT,
+    SUCCESS_COLOR,
+    WARNING_COLOR,
+)
+
+RUNNING_COLOR = '#008000'
+STOPPED_COLOR = '#ffff00'
+
+__all__ = [
+    'DANGER_COLOR',
+    'INFO_COLOR',
+    'PRIMARY_COLOR',
+    'RUNNING_COLOR',
+    'SECONDARY_COLOR',
+    'SECONDARY_COLOR_LIGHT',
+    'STOPPED_COLOR',
+    'SUCCESS_COLOR',
+    'WARNING_COLOR',
+]