+++ ubo_app/services/050-vscode/commands.py
@@ -9,8 +9,8 @@ from typing import Literal, TypedDict
 
 from _constants import CODE_BINARY_PATH
 from debouncer import DebounceOptions, debounce
-from ubo_gui.constants import DANGER_COLOR
 
+from ubo_app.colors import DANGER_COLOR
 from ubo_app.logger import logger
 from ubo_app.store.main import store
 from ubo_app.store.services.notifications import (